sources: Transmit errors from sub command as JSON
Before, OSBuild was transmitting errors about sources download as
RuntimeErrors, transmitted as raw python stack to its caller.

Now OSBuild transmit the error as json objects that can be parsed by a
machine (i.e Composer). Each error has an ID associated, brings a human
readable message and an optional payload. With this information the
caller can reconstruct what happened and transmit that up the stack.

Ultimate goal is to bring feedback to the final user with the most
simple example being:
"download of {url} failed because {reason}"